==> Notebooks/10_scrapping.ipynb <==
%load_ext autoreload
%autoreload 2
%load_ext dotenv
%dotenv

from dotenv import load_dotenv
import os
import pandas as pd
import yaml, json, time
from src.data_fetcher import (
    fetch_weather_data,
    fetch_energy_data,
)



load_dotenv()          # Load environment variables from .env file  

token = os.environ.get("NAOO_TOKEN")
eia_token = os.environ.get("EIA_KEY")

if token is None or eia_token is None:
    print("Error retrieving token") 
else:
    print("Token retrieved successfully")



with open("../config/config.yaml") as f:
    data = yaml.safe_load(f)

cities_list = data["cities"]

print(cities_list)

output_file = "../data/raw/all_weather.csv"
first = True    #write header only for the first file

for city in cities_list:
    print(f"Fetching data for {city["city"]}...")
    
    result = fetch_weather_data(city["station"], "2025-03-01", "2025-07-26")
    df = pd.DataFrame(result["results"])
    df["city"] = city["city"]
    df["state"] = city["state"]
    
    # Write to CSV with mode (w) at the first loop, then (a) for subsequent loops
    df.to_csv(output_file,
              mode = "w" if first else "a",
              header = first,
              index = False)
    first = False
    
    print(f"Saved {len(df)} rows for {city['city']}. Sleeping for 10 Seconds…")
    print()
    print("-"*50)
    time.sleep(10)

# fetch initial energy data

print(cities_list)

the idea is to extract the demand and the net generation seperately and merge both data

output_file = "../data/raw/all_energy.csv"
first = True    #write header only for the first file
data = None

for city in cities_list:
    print(f"Fetching data for {city["city"]}...")
    
    region = city["region"]
    timezone = city["timezone"]
    start = "2025-03-01"
    end = "2025-07-26"
    types = ["D", "NG"]
    for type in types:
        energy_result = fetch_energy_data(region=region, types=type, timezone=timezone, start=start, end=end)
        data = energy_result["response"]["data"]
        


        df = pd.DataFrame(data)
        df["city"] = city["city"]
        df["state"] = city["state"]
    
        # Write to CSV with mode (w) at the first loop, then (a) for subsequent loops
        df.to_csv(output_file,
            mode = "w" if first else "a",
            header = first,
            index = False)
        first = False
    
        print(f"Saved {len(df)} rows for {city['city']} and {type}. Sleeping for 10 Seconds…")
        print()
    print("-"*50)
    time.sleep(10)
