==> Notebooks/20_exploration.ipynb <==
Exploring temperature data

%load_ext autoreload
%autoreload 2
%load_ext dotenv
%dotenv

import pandas as pd

df_temp = pd.read_csv("../data/raw/all_weather.csv")
df_temp.head()

weather_pivot = df_temp.pivot_table(
    index=["date", "city", "state"],
    columns="datatype",
    values="value"
).reset_index()


weather_pivot.head(10)

Exploring energy data

energy_df = pd.read_csv("../data/raw/all_energy.csv")
energy_df.head(2)

energy_pivot = energy_df.pivot_table(
    index= ['period', 'respondent-name', "timezone", "city", "state", "value-units"],
    columns=["type-name"],
    values="value"
).reset_index()

# Convert both to datetime
weather_pivot["date"] = pd.to_datetime(weather_pivot["date"]).dt.date
energy_pivot["period"] = pd.to_datetime(energy_pivot["period"]).dt.date

# rename period to date
energy_pivot = energy_pivot.rename(columns={'period':'date'})
energy_pivot.head()

# Merge on date, city, and state
merged_df = pd.merge(weather_pivot, energy_pivot, on=["date", "city", "state"], how="inner")

merged_df.head(20)

merged_df.to_csv("../data/processed/weather_energy_data.csv", index=False)


from src.data_fetcher import (
    get_temp_last_date,
    get_energy_last_date,
    get_cities
)

# get temp

start_temp = get_temp_last_date("../data/raw/all_weather.csv")
print(start_temp)
print(type(start_temp))

# get enrgy

start_energy = get_energy_last_date("../data/raw/all_energy.csv")
print(start_energy)
print(type(start_energy))

city = get_cities("../config/config.yaml")

for i in city:
    print(i)

import datetime
print(datetime.date.today() + datetime.timedelta(days=1))

if start_energy > datetime.date.today():
    print("hello")
elif start_energy < datetime.date.today():
    print("world")
else:
    print("none")

import pandas as pd
from datetime import datetime, timedelta

# 1. Load existing CSV
weather_path = "../data/raw/all_weather.csv"
df = pd.read_csv(weather_path, parse_dates=["date"])
df.head(2)

df["date"] = pd.to_datetime(df["date"])

# 2. Get last date per city
last_dates = df.groupby("city")["date"].max().to_dict()
print(last_dates)



for key, value in 


import yaml
with open("../config/config.yaml") as f:
    data = yaml.safe_load(f)

cities_list = data["cities"]

# print(cities_list)

# 3. Use it to fetch data for each city
for city_info in cities_list:  # your YAML cities
    city = city_info["city"]

    
    # Use last saved date + 1 day or a default start
    start_date = last_dates.get(city, pd.Timestamp("2023-01-01")) + timedelta(days=1)
    # end_date = datetime.today()
    print(city, start_date)
